# customer-service/src/main/resources/openapi/CustomerService.yaml
openapi: 3.0.1
info:
  title: Customer Service
  version: v1
security:
  - BearerAuth: []

tags:
  - name: Customers
    description: API for working with clients

paths:
  /api/v1/customers/{id}:
    put:
      tags:
        - Customers
      summary: Update customer details
      operationId: updateCustomer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequestDto'
        required: true
      responses:
        '200':
          description: Customer details have been updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponseDto'
        '400':
          description: Incorrect request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                ValidationError:
                  $ref: '#/components/examples/ValidationErrorExample'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                Unauthorized:
                  $ref: '#/components/examples/UnauthorizedExample'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                NotFound:
                  $ref: '#/components/examples/NotFoundExample'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                InternalServerError:
                  $ref: '#/components/examples/InternalServerErrorExample'

  /api/v1/customers:
    get:
      tags:
        - Customers
      summary: Get a list of clients
      operationId: getAllCustomers
      parameters:
        - name: fullName
          in: query
          required: false
          schema:
            type: string
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Client list
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Page'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                Unauthorized:
                  $ref: '#/components/examples/UnauthorizedExample'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                InternalServerError:
                  $ref: '#/components/examples/InternalServerErrorExample'
    post:
      tags:
        - Customers
      summary: Create new Customer
      operationId: createCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequestDto'
        required: true
      responses:
        '200':
          description: Client created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomerResponseDto'
        '400':
          description: Incorrect request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                ValidationError:
                  $ref: '#/components/examples/ValidationErrorExample'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                Unauthorized:
                  $ref: '#/components/examples/UnauthorizedExample'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                InternalServerError:
                  $ref: '#/components/examples/InternalServerErrorExample'

  /api/v1/customers/{guid}:
    get:
      tags:
        - Customers
      summary: Find customer by ID
      description: Returns information about the client
      operationId: Id
      parameters:
        - name: guid
          in: path
          description: Unique identifier of the customer
          required: true
          schema:
            type: string
          example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        '200':
          description: Client found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CustomerResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                Unauthorized:
                  $ref: '#/components/examples/UnauthorizedExample'
        '404':
          description: Client not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                NotFound:
                  $ref: '#/components/examples/NotFoundExample'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
              examples:
                InternalServerError:
                  $ref: '#/components/examples/InternalServerErrorExample'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  examples:
    ValidationErrorExample:
      value:
        timestamp: '2025-08-12T10:15:30Z'
        status: 400
        error: "Bad Request"
        message: "Validation failed: Email format is invalid"
        path: "/api/v1/customers"

    UnauthorizedExample:
      value:
        timestamp: '2025-08-12T10:15:30Z'
        status: 401
        error: "Unauthorized"
        message: "Full authentication is required to access this resource"
        path: "/api/v1/customers"

    NotFoundExample:
      value:
        timestamp: '2025-08-12T10:15:30Z'
        status: 404
        error: "Not Found"
        message: "Customer with ID 123e4567-e89b-12d3-a456-426614174000 not found"
        path: "/api/v1/customers/123e4567-e89b-12d3-a456-426614174000"

    InternalServerErrorExample:
      value:
        timestamp: '2025-08-12T10:15:30Z'
        status: 500
        error: "Internal Server Error"
        message: "Unexpected error occurred while processing the request"
        path: "/api/v1/customers"

  schemas:
    ErrorDto:
      type: object
      required:
        - status
        - message
        - timestamp
      properties:
        timestamp:
          type: string
          format: date-time
          description: Time when the error occurred
        status:
          type: integer
          format: int32
          description: HTTP status code
        error:
          type: string
          description: Short error title
        message:
          type: string
          description: Detailed error description
        path:
          type: string
          description: The request URI that caused the error

    UpdateCustomerRequestDto:
      required:
        - email
        - name
      type: object
      properties:
        name:
          maxLength: 30
          minLength: 1
          type: string
          description: Client fullname
        email:
          pattern: '[^@ \t\r\n]+@[^@ \t\r\n]+\.[^@ \t\r\n]+'
          type: string
          description: Client's email
          example: customer@example.com
        phoneNumber:
          pattern: ^(\+\d{1,2}\s?)?\(?\d{3}\)?[\s.-]?\d{3}[\s.-]?\d{4}$
          type: string
          description: Phone number
          example: "+375295647594"
      description: Client update request
    ContactDetailsResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Contact details ID
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          description: Client's email
          example: customer@example.com
        phoneNumber:
          pattern: ^(\+\d{1,2}\s?)?\(?\d{3}\)?[\s.-]?\d{3}[\s.-]?\d{4}$
          type: string
          description: Phone number
        createdAt:
          type: string
          description: Date of record creation
          format: date-time
          example: '2025-08-11T12:34:56Z'
        updatedAt:
          type: string
          description: Date of last update of the entry
          format: date-time
          example: '2025-08-12T15:00:00Z'
      description: Client contact details
    CustomerResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Client ID
          example: 220e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          description: Client full name
          example: Ivan Ivanov
        contactDetails:
          $ref: '#/components/schemas/ContactDetailsResponseDto'
        createdAt:
          type: string
          description: Client creation date
          format: date-time
          example: '2025-08-11T12:34:56Z'
        updatedAt:
          type: string
          description: Date of last client update
          format: date-time
          example: '2025-08-12T15:00:00Z'
      description: Client information
    CreateCustomerRequestDto:
      required:
        - email
        - name
      type: object
      properties:
        name:
          maxLength: 30
          minLength: 1
          type: string
          description: Client fullname
        email:
          pattern: '[^@ \t\r\n]+@[^@ \t\r\n]+\.[^@ \t\r\n]+'
          type: string
          description: Client's email
        phoneNumber:
          pattern: ^(\+\d{1,2}\s?)?\(?\d{3}\)?[\s.-]?\d{3}[\s.-]?\d{4}$
          type: string
          description: Phone number
      description: Request to create a client
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    Page:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            type: object
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean